spark_processing/requirements.txt
pyspark
cassandra-driver
